requirements.txt
fastapi
uvicorn[standard]
openai
langchain
faiss-cpu
tiktoken
transformers
pydantic
python-dotenv
biopython